refactor(hpm): share MCAN CAN backend
