+++ driver/hpm/hpm_can.hpp
@@ -1,22 +1,14 @@
 #pragma once
 
-#include "hpm_clock_drv.h"
-#include "hpm_soc.h"
+#include "hpm_mcan_backend.hpp"
 
-#if __has_include("hpm_mcan_drv.h") && __has_include("hpm_mcan_regs.h") && \
-                                                     __has_include("hpm_mcan_soc.h")
-#include "hpm_mcan_drv.h"
 // 当前 HPM CAN 支持假定 MCAN message RAM 位置由
 // `MCAN_MSG_BUF_BASE_VALID_START` 描述。
 // Current HPM CAN support assumes MCAN message RAM placement is described by
 // `MCAN_MSG_BUF_BASE_VALID_START`.
-#if ((defined(MCAN_SOC_MAX_COUNT) && (MCAN_SOC_MAX_COUNT > 0)) || \
-     (defined(CAN_SOC_MAX_COUNT) && (CAN_SOC_MAX_COUNT > 0))) &&  \
-    defined(HPM_MCAN0) && defined(MCAN_MSG_BUF_BASE_VALID_START)
+#if LIBXR_HPM_MCAN_SUPPORTED && defined(HPM_MCAN0) && \
+    defined(MCAN_MSG_BUF_BASE_VALID_START)
 #define LIBXR_HPM_CAN_SUPPORTED 1
-#else
-#define LIBXR_HPM_CAN_SUPPORTED 0
-#endif
 #else
 #define LIBXR_HPM_CAN_SUPPORTED 0
 #endif
@@ -25,9 +17,7 @@
 
 #include <atomic>
 #include <cstdint>
-#include <cstring>
 
-#include "can.hpp"
 #include "queue.hpp"
 
 namespace LibXR
@@ -56,11 +46,8 @@ class HPMCAN final : public CAN
   ErrorCode SetMessageBuffer(void* msg_buf, uint32_t msg_buf_size);
 
  private:
-  static ErrorCode ConvertStatus(hpm_stat_t status);
-  static mcan_node_mode_t ConvertMode(const CAN::Mode& mode);
   static void BuildTxFrame(const ClassicPack& pack, mcan_tx_frame_t& frame);
   static void BuildRxPack(const mcan_rx_message_t& frame, ClassicPack& pack);
-  static ErrorID ConvertLastError(uint8_t last_error);
 
   void EnableCanInterrupts();
   void DisableCanInterrupts();
